+++ CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Training session-specific exercise notes
+- 1RM calculator accessible from the navigation menu and from the training session exercise options
 - Routine page: Option to show and copy routine as text
 - Routines page: Option to show and copy routine as text
+- Exercise page: 1RM chart
 - Body weight, body fat, exercises and routines pages: Automatic focus of the first input field in dialogs for adding entries
 - CLI commands to list, create, update and delete users